books_database/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
watchdog